refactor(exporter): unify loader pipeline + extract flat_scanner.go (PR-7)
Final piece of the v2.7.0+ hot-reload code-health story. Pure
structural refactor — zero behavior change, race x10 + benchmark
within ±4% noise.

## Loader pipeline unification

Three loaders (Load + IncrementalLoad + fullDirLoad) all duplicated
the "atomic-swap + post-commit hooks" tail. PR-7 extracts:

  commitConfig(cfg, hash, flatScan, logHeader)
    — under m.mu.Lock: install config + loaded + lastReload + lastHash
      + (optionally) flat scan state. Then detectConfigSource +
      logConfigStats. Single seam for ALL three loaders' atomic swap.

  runHierarchyScanReject(label)
    — populateHierarchyState with the consistent error policy:
      *DuplicateTenantError → propagate as hard reject (issue #127);
      generic errors → log WARN, return nil (hierarchical mode is
      opt-in, malformed branch shouldn't tear down flat-only deploy).
      label appears in the WARN line for ops triage.

Load (single-file + dir): 65 LoC → 35 LoC after pulling commit + reject
fullDirLoad: 80 LoC → 65 LoC
IncrementalLoad atomic-swap tail: 13 LoC → 6 LoC

Net config.go: 1185 → 783 LoC (-34%) after also moving flat helpers
to flat_scanner.go (below).

## flat_scanner.go extraction (audit item #16)

Moved 5 flat-mode helpers from the giant config.go into a dedicated
flat_scanner.go (371 LoC) that lives next to PR-5's flatScanState:

  loadFile(path)            — single YAML → ThresholdConfig + hash
  loadDir(dir)              — dir-mode merge (used by Load)
  scanDirFileHashes(...)    — per-file SHA-256 + mtime fast-path
                              (used by IncrementalLoad + fullDirLoad)
  applyBoundaryRules(...)   — _-prefix vs tenant file convention
  mergePartialConfigs(...)  — multi-file partial → ThresholdConfig

Imports trimmed: config.go drops `crypto/sha256` (only used by the
flat helpers). flat_scanner.go imports the union of what its 5
functions need.

## Verification

  - go build -buildvcs=false ./... clean
  - go vet ./... clean
  - Linux dev container worktree path:
    `go test -race -count=1 ./...` — 9/9 packages pass (5.4s)
  - Hot-path stress -race -count=10 (TestSlowWriteTornStateStress +
    TestDebounce + TestDiffAndReload + TestIncrementalLoad +
    TestWatchLoop + TestComputeMergedHash + TestGoldenParity +
    TestConfigManager_Load) — 39.3s, 0 races, 0 flakes.
  - Apples-to-apples bench (Windows host, parallel git worktree
    discipline learned from PR-4):
      IncrementalLoad_1000_OneFileChanged: 33.9ms → 35.2ms (+3.9%)
      DiffAndReload_Hierarchical_1000_NoChange: 393ms → 394ms (+0.3%)
      FullDirLoad_Hierarchical_2000: 645ms → 628ms (-2.6%)
      DiffAndReload_Hierarchical_2000_NoChange: 760ms → 752ms (-1.1%)
    All within ±4% noise envelope.

## v2.8.0 follow-up sequence

  PR-6 (#199) ✅ merged — emit.go 3-way split
  PR-7 (this) ⏳ open    — loader unify + flat_scanner extract
  PR-8 next              — pkg/config library maturation
  PR-9 next              — test infra modernization
